Moved MotorControlTimers from compose package to the utils package
