+++ MotorControl/src/main/kotlin/org/sagebionetworks/motorcontrol/state/ActiveStep.kt
@@ -19,7 +19,7 @@ import kotlinx.coroutines.launch
 import org.sagebionetworks.assessmentmodel.Result
 import org.sagebionetworks.assessmentmodel.passivedata.recorder.motion.MotionRecorderConfiguration
 import org.sagebionetworks.motorcontrol.navigation.HandSelection
-import org.sagebionetworks.motorcontrol.presentation.compose.StepTimer
+import org.sagebionetworks.motorcontrol.utils.StepTimer
 import org.sagebionetworks.motorcontrol.recorder.RecorderRunner
 import org.sagebionetworks.motorcontrol.recorder.RecorderScheduledAssessmentConfig
 import org.sagebionetworks.motorcontrol.serialization.BackgroundRecordersConfigurationElement
